@@ -15,6 +15,17 @@ export default class AssetUtils {
             committee_fed_asset  : false
         }
 
+        if (mask === "all") {
+            for (let flag in booleans) {
+                if (!isBitAsset && (assetConstants.uia_permission_mask.indexOf(flag) === -1)) {
+                    delete booleans[flag];
+                } else {
+                    booleans[flag] = true;
+                }
+            }
+            return booleans;
+        }
+
         for (let flag in booleans) {
             if (!isBitAsset && (assetConstants.uia_permission_mask.indexOf(flag) === -1)) {
                 delete booleans[flag];
@@ -46,11 +57,26 @@ export default class AssetUtils {
         let permissions = isBitAsset ? Object.keys(assetConstants.permission_flags) : assetConstants.uia_permission_mask;
         let flags = 0;
         permissions.forEach(permission => {
-            if (flagBooleans[permission]) {
+            if (flagBooleans[permission] && permission !== "global_settle") {
                 flags += assetConstants.permission_flags[permission];
             }
         })
 
+        if (isBitAsset) {
+            flags += assetConstants.permission_flags["global_settle"];
+        }
+
         return flags;
+    }
+
+    static parseDescription(description) {
+        let parsed;
+        try {
+            parsed = JSON.parse(description)
+        } catch (error) {
+            
+        }
+
+        return parsed ? parsed : {main: description};
     }
 }
