@@ -15,11 +15,11 @@
         dashboard: "Hesaplar",
         explorer: "Araştır",
         exchange: "Al / Sat",
-        payments: "Transferler", 
+        payments: "Transfer", 
         logout: "Çıkış",
         settings: "Ayarlar",
         current: "Kullanılan Hesap",
-        create_account: "Hesap Oluştur",
+        create_account: "Oluştur",
         create_asset: "Aktif Oluştur",
         update_asset: "Aktif Güncelle",
         lock: "Kiliti Kapat",
@@ -31,7 +31,7 @@
     propose: "teklif et",
     cancel: "iptal et",
     account: {
-        welcome: "Graphene'e Hoşgeldiniz",
+        welcome: "Blokzinciri'ne Hoşgeldiniz",
         asset: "Aktif",
         market_value: "Piyasa Değeri",
         hour_24: "24 Saat",
@@ -39,30 +39,30 @@
         name: "Hesap adı",
         id: "Hesap Kimlik",
         more: "fazlası",
-        deposit_withdraw: "Yatır/Çek",
+        deposit_withdraw: "Yatır / Çek",
         as_collateral: "Teminat",
         open_orders: "Açık emirler",
         total_value: "Toplam değeri",
-        show_hidden: "Saklı aktifleri göster",
-        hide_hidden: "Saklı aktifleri gizle",
+        show_hidden: "Göster",
+        hide_hidden: "Gizle",
         ignore: "Yok say",
         unignore: "Var say",
-        show_ignored: "Yok sayılanı göster",
-        hide_ignored: "Yok sayılanı gizle",
+        show_ignored: "Göster",
+        hide_ignored: "Gizle",
         whitelist: {
-            title: "Beyazliste",
-            black: "Karaliste",
+            title: "Beyaz Liste",
+            black: "Kara Liste",
             add: "Beyaz listeye ekle",
-            add_black: "Karalisteye ekle",
-            empty: "%(account)s herhangi bir hesabı beyazlisteye eklemedi.",
-            empty_black: "%(account)s herhangi bir hesabı karalisteye eklemedi.",
-            white_by: "Beyazlisteye alındı",
-            black_by: "Karalisteye alındı",
-            empty_white_by: "%(account)s adlı hesabı beyazlisteye alan olmadı.",
-            empty_black_by: "%(account)s adlı hesabı karalisteye alan olmadı."
+            add_black: "Kara listeye ekle",
+            empty: "%(account)s herhangi bir hesabı beyaz listeye eklemedi.",
+            empty_black: "%(account)s herhangi bir hesabı kara listeye eklemedi.",
+            white_by: "Beyaz listeye alan",
+            black_by: "Kara listeye alan",
+            empty_white_by: "%(account)s adlı hesabı beyaz listeye alan olmadı.",
+            empty_black_by: "%(account)s adlı hesabı kara listeye alan olmadı."
         },
         vesting: {
-            title: "Tutulan bakiyeler",
+            title: "Tutulan Bakiyeler",
             balance_number: "Bakiye #%(id)s",
             no_balances: "Bu hesaba ait tutulan bir bakiye yok"
         },
@@ -71,11 +71,11 @@
             join: "Katılma tarihi",
             registrar: "Kaydı yapan",
             referrer: "Yeni Üye Referrer",
-            lifetime_referrer: "Ömür-boyu Referrer",
+            lifetime_referrer: "Ömür-Boyu Referrer",
             network_percentage: "Ağ",
             fee_allocation: "Ücret Tahsisi",
             membership: "Üyelik",
-            fees_paid: "Ödenmiş ömür-boyu ücretler",
+            fees_paid: "Ödenmiş Ömür-Boyu ücretler",
             fees_pending: "Bekleyen ücretler",
             fees_vested: "Bekleyen hak ücretler",
             referrals: "Referraller",
@@ -123,9 +123,9 @@
             new_issuer: "Yeni sahip hesabı",
             flags: "Bayraklar",
             charge_market_fee: "Piyasa ücretinin etkin kıl",
-            override_authority: "Aktifi ihraç eden geri kendine transfer edebilir",
+            override_authority: "Aktifi geri transfer edebilir",
             transfer_restricted: "İhrac eden tüm transferleri onlaylamak zorunda",
-            disable_confidential: "Özel işlemleri etkisizleştir",
+            disable_confidential: "Gizli işlemler etkisiz",
             white_list: "Tutanları beyaz-listelemeyi gerektir",
             witness_fed_asset: "Tanıkların fiyat beslemelerine müsade et",
             committee_fed_asset: "Kurul üyelerinin fiyat beslemelerine müsade et",
@@ -134,11 +134,26 @@
             advanced: "Gelişmiş",
             need_asset: " İki aktiften biri an azından %(name)s olmalıdır",
             perm_warning: "UYARI: İzinler yalnızca etkisiz hale getirilebilitler, bir kere etkisiziştiridiklerinde tekrar aktif hale getirilemezler!!",
-            issued_assets: "İhraç Edilmiş Aktifler"
+            issued_assets: "Piyasa Sürülen Aktifler",
+            short: "Kısa ad (max 32 harf)",
+            condition: "Durum (max 60 harf)",
+            expiry: "Piyasa çözülüm tarihi",
+            decimals: "Ondalık hane sayısı",
+            mpa: "Akıllı Para",
+            pm: "İki değerli tahmin piyasası",
+            bitasset_opts: "Akıllı Para tercihleri",
+            feed_lifetime_sec: "Dakika olarak yayın ömrü",
+            minimum_feeds: "En az yayın sayısı",
+            force_settlement_delay_sec: "Zorunlu hesaplaşma gecikmesi (dakika)",
+            force_settlement_offset_percent: "Zorunlu hesaplaşmanın yüzdelik sapması",
+            maximum_force_settlement_volume: "Max zorunlu mahsup hacmi (yüzde)",
+            backing: "Kısanın dayandığı aktif",
+            error_precision: "Aktif %(asset)s ile aynı hassasiyete sahip değil",
+            error_invalid: "O aktif kullanılamaz"
         },
         connections: {
             known: "Tarafından Tanınıyor",
-            "black": "Tarafından Karalistelenmiş"
+            "black": "Tarafından Kara listelenmiş"
         },
         perm: {
             active: "Etkin İzinler",
@@ -182,8 +197,8 @@
             add_committee: "Ekle",
             add_committee_label: "Kurul Üyesi",
             add_witness_label: "Tanık",
-            approve_worker: "Onayla",
-            reject_worker: "Reddet",
+            approve_worker: "Kabul",
+            reject_worker: "Red",
             worker_account: "Emekçi hesabı",
             total_votes: "Net oylar",
             votes_against: "Aleyhine oylar",
@@ -219,7 +234,7 @@
         history: "Geçmiş",
         payees: "Alıcılar",
         permissions: "İzinler",
-        voting: "Oylama",
+        voting: "Seçimler",
         orders: "Emirler",
         select_placeholder: "Hesap Seç...",
         errors: {
@@ -231,7 +246,7 @@
         collaterals: "Teminat Pozisyonları",
         eq_value: "Eş Değeri",
         percent: "Toplam arzın yüzdesi",
-        please_create_account: "Lütfen bir hesap açın",
+        please_create_account: "Lütfen bir hesap oluşturun",
         create_account: "Hesap oluştur",
         identicon: "Identicon",
         pay_from: "Buradan öde",
@@ -302,9 +317,9 @@
         asset_update: "{account} şu aktifi güncelledi: {asset}",
         lifetime_upgrade_account: "{account} ömür-boyu üyeliğe yükseltildi",
         annual_upgrade_account: "{account} yıllık üyeliğe yükseltildi",
-        unlisted_by: "{lister}  {listee} adlı hesabı listeden çıkardı",
-        whitelisted_by: "{lister}  {listee} adlı hesabı beyaz listeye aldı",
-        blacklisted_by: "{lister}  {listee} adlı hesabı kara-listeye aldı",
+        unlisted_by: "{lister} listeden {listee} adlı hesabı çıkardı",
+        whitelisted_by: "{lister} tarafından {listee} beyaz listeye alındı",
+        blacklisted_by: "{lister} tarafından {listee} kara-listeye aldı",
         transfer_account: "{account} adlı hesabın mülkiyeti {to} adlı hesaba devredildi.",
         asset_update_feed_producers: "{account} şu aktifin fiyat kaynaklarını güncelledi:{asset}",
         asset_fund_fee_pool: "{account} tarafından {asset} ücret havuzuna {amount} eklendi",
@@ -390,8 +405,8 @@
             account_whitelist: "Beyaz Listeye Alma",
             account_upgrade: "Hesap yükseltme",
             account_transfer: "Hesap transferi",
-            asset_create: "Aktif oluşturulması",
-            asset_update: "Aktif güncellemesi",
+            asset_create: "Aktif oluştur",
+            asset_update: "Aktif güncelle",
             asset_update_bitasset: "Akıllıpara güncelleme",
             asset_update_feed_producers: "Aktif feed üreticileri güncelleme",
             asset_issue: "Aktif düzenle",
@@ -440,16 +455,16 @@
             basic_fee : "Temel Ücret",
             premium_fee : "Paralı Adlar İçin Ücret",
             membership_annual_fee : "Yıllık Üyelik",
-            membership_lifetime_fee : "Ömür-boyu Üyelik",
+            membership_lifetime_fee : "Ömür-Boyu Üyelik",
             symbol3 : "3 Haneli Semboller",
             symbol4 : "4 Haneli Semboller",
             long_symbol : "Uzun Semboller"
         },
         whitelist_states: {
             no_listing: "Listeleye alınmamış",
-            white_listed: "Beyazlisteye alınmış",
-            black_listed: "Karalisteye alınmış",
-            white_and_black_listed: "Beyazlisteye ve karalisteye alınmış"
+            white_listed: "Beyaz listeye alınmış",
+            black_listed: "Kara listeye alınmış",
+            white_and_black_listed: "Beyaz ve kara listeye alınmış"
         }
     },
     explorer: {
@@ -507,8 +522,9 @@
         },
         assets: {
             title: "Aktifler",
-            market: "AkıllıParalar",
+            market: "Akıllı Paralar",
             user: "Kullanıcı Aktifleri",
+            prediction: "Tahmin piyasası",
             symbol: "Sembol",
             id: "Kimlik",
             issuer: "İhraç Eden",
@@ -544,8 +560,8 @@
                 title: "İzinler",
                 max_market_fee: "Max piyasa ücreti",
                 max_supply: "Max arz",
-                blacklist_authorities: "Karaliste yetkiler",
-                blacklist_markets: "Karaliste piyasları",
+                blacklist_authorities: "Kara liste yetkiler",
+                blacklist_markets: "Kara liste piyasları",
                 whitelist_authorities: "Beyaz Liste yetkiler",
                 whitelist_markets: "Beyaz Liste piyasaları"
             },
@@ -653,7 +669,7 @@
         market_name: "Piyasalarım",
         quote_supply: "Karşıt arzı",
         base_supply: "Taban arzı",
-        more: "Tüm piyasalar",
+        more: "Piyasalar bul",
         volume_24: "24 Saat",
         change: "Değişim",
         confirm: "Verdiğiniz emir şimdiki fiyattan %(diff)s% kadar farklı, emin misiniz?",
@@ -668,6 +684,7 @@
         index: "Indeks",
         my_bids: "Alış Fiyatlarım",
         my_asks: "Satış Fiyatlarım",
+        my_orders: "Emirlerim",
         settle_orders: "Emirleri öde",
         asks: "Satış Fiyatları",
         bids: "Alış Fiyatları",
@@ -676,7 +693,11 @@
         borrow: "Ödünç Al",
         no_balance: "Yetersiz bakiye",
         invalid_amount: "Geçersiz miktar",
-        invalid_price: "Geçersiz fiyat"
+        invalid_price: "Geçersiz fiyat",
+        show_bids: "Tüm alış fiyat göster",
+        show_asks: "Tümünü göster",
+        hide: "Gizle",
+        short: "Kısa"
     },
     fees: {
      title: "Ücret Tarifesi"
@@ -693,7 +714,7 @@
     },
     wallet: {
         title: "Cüzdan",
-        confirm: "Parola (onayla)",
+        confirm: "Parola (teyit)",
         password: "Parola",
         existing_password: "Mevcut parola",
         change_password: "Parolayı Değiştir",
@@ -795,6 +816,12 @@
         title: "%(asset)s ödenmesini talep et",
         amount: "Ödenecek miktar",
         submit: "Aktif tasfiye et"
+      },
+      reserve: {
+        title: "Rezerv (yakılan) aktif",
+        amount: "Rezerv edilecek miktar",
+        from: "Rezerv yapılacak hesap",
+        submit: "Rezerv"
       },
       ok: "Tamam"
     },
