@@ -8,6 +8,7 @@ import MarketHistory from "./MarketHistory";
 import MyMarkets from "./MyMarkets";
 import BuySell from "./BuySell";
 import utils from "common/utils";
+import assetUtils from "common/asset_utils";
 import PriceChart from "./PriceChart";
 import DepthHighChart from "./DepthHighChart";
 import {debounce, cloneDeep} from "lodash";
@@ -36,7 +37,8 @@ import OpenSettleOrders from "./OpenSettleOrders";
 require("./exchange.scss");
 
 let emitter = ee.emitter();
-let callListener, limitListener, newCallListener;
+let callListener, limitListener, newCallListener, feedUpdateListener, settleOrderListener;
+let SATOSHI = 8;
 
 Highcharts.setOptions({
     global: {
@@ -102,7 +104,10 @@ class Exchange extends React.Component {
         this.state = this._initialState(props);
     }
 
+
     _initialState(props) {
+        let ws = props.viewSettings;
+
         return {
             history: [],
             buyAmount: 0,
@@ -132,19 +137,21 @@ class Exchange extends React.Component {
             sellAmount: 0,
             sellTotal: 0,
             sub: null,
-            flipBuySell: props.viewSettings.get("flipBuySell"),
+            flipBuySell: ws.get("flipBuySell"),
             favorite: false,
-            showDepthChart: props.viewSettings.get("showDepthChart"),
-            leftOrderBook: props.viewSettings.get("leftOrderBook"),
+            showDepthChart: ws.get("showDepthChart"),
+            leftOrderBook: ws.get("leftOrderBook"),
             buyDiff: false,
             sellDiff: false,
-            indicators: props.viewSettings.get("indicators") || {
+            indicators: ws.get("indicators") || {
                 rsi: false,
                 sma: false,
                 atr: false,
                 ema: false
             },
-            indicatorSettings: props.viewSettings.get("indicatorSettings") || {
+            preferCoreBuyFee: ws.get("preferCoreBuyFee") !== undefined ? ws.get("preferCoreBuyFee") : true,
+            preferCoreSellFee: ws.get("preferCoreSellFee") !== undefined ? ws.get("preferCoreSellFee") : true,
+            indicatorSettings: ws.get("indicatorSettings") || {
                 rsi: {
                     period: 14,
                     overbought: 70,
@@ -206,6 +213,15 @@ class Exchange extends React.Component {
         emitter.on('cancel-order', limitListener = MarketsActions.cancelLimitOrderSuccess);
         emitter.on('close-call', callListener = MarketsActions.closeCallOrderSuccess);
         emitter.on('call-order-update', newCallListener = MarketsActions.callOrderUpdate);
+        emitter.on('bitasset-update', feedUpdateListener = MarketsActions.feedUpdate);
+        emitter.on('settle-order-update', settleOrderListener = (object) => {
+            let {isMarketAsset, marketAsset} = market_utils.isMarketAsset(this.props.quoteAsset, this.props.baseAsset);
+            console.log("settle-order-update:", object, "isMarketAsset:", isMarketAsset, "marketAsset:", marketAsset);
+
+            if (isMarketAsset && marketAsset.id === object.balance.asset_id) {
+               MarketsActions.settleOrderUpdate(marketAsset.id);
+            }
+        });
     }
 
     componentDidMount() {
@@ -226,6 +242,12 @@ class Exchange extends React.Component {
     }
 
     componentWillReceiveProps(nextProps) {
+        if (nextProps.baseAsset && nextProps.baseAsset.getIn(["bitasset", "is_prediction_market"])) {
+            console.log(nextProps.baseAsset.get("symbol"), "is prediction market");
+            // console.log("this.props:", this.props);
+            this.props.history.push(`market/${nextProps.baseAsset.get("symbol")}_${nextProps.quoteAsset.get("symbol")}`)
+        }
+
         if (nextProps.quoteAsset.toJS && nextProps.baseAsset.toJS) {
             // this._addMarket(nextProps.quoteAsset.get("symbol"), nextProps.baseAsset.get("symbol"));
             if (!this.state.sub) {
@@ -251,18 +273,21 @@ class Exchange extends React.Component {
         emitter.off('cancel-order', limitListener);
         emitter.off('close-call', callListener);
         emitter.off('call-order-update', newCallListener);
+        emitter.off('bitasset-update', feedUpdateListener);
     }
 
-    _createLimitOrder(buyAsset, sellAsset, buyAssetAmount, sellAssetAmount, feeID) {
-        console.log("createLimitOrder:", buyAssetAmount, sellAssetAmount);
+    _createPredictionShort(buyAsset, sellAsset, buyAssetAmount, sellAssetAmount, feeID) {
+        console.log("createPredictionShort:", buyAssetAmount, sellAssetAmount);
+        let coreAsset = ChainStore.getAsset("1.3.0");
         let expiration = new Date();
         // TODO: Add selector for expiry
         expiration.setYear(expiration.getFullYear() + 5);
-        MarketsActions.createLimitOrder(
+        MarketsActions.createPredictionShort(
             this.props.currentAccount.get("id"),
             utils.get_satoshi_amount(sellAssetAmount, sellAsset),
             sellAsset,
             utils.get_satoshi_amount(buyAssetAmount, buyAsset),
+            utils.get_satoshi_amount(sellAssetAmount, coreAsset),
             buyAsset,
             expiration,
             false, // fill or kill TODO: add fill or kill switch
@@ -281,7 +306,7 @@ class Exchange extends React.Component {
     _getFee(asset) {
         let fee = utils.estimateFee("limit_order_create", [], ChainStore.getObject("2.0.0")) || 0;
 
-        if (!asset) return fee;
+        if (!asset || asset.get("id") === "1.3.0") return fee;
         let cer = asset.getIn(["options", "core_exchange_rate"]).toJS();
         let coreAsset = ChainStore.getAsset("1.3.0");
         if (!coreAsset) return 0;
@@ -292,23 +317,30 @@ class Exchange extends React.Component {
         return eqValue;
     }
 
-    _verifyFee(fee, feeAsset, sellAmount, sellBalance, coreBalance) {
+    _verifyFee(fee, feeAsset, sellAsset, sellAmount, sellBalance, coreBalance) {
         let coreFee = this._getFee();
 
-        let sellPrecision = utils.get_asset_precision(feeAsset);
+        let sellPrecision = utils.get_asset_precision(sellAsset);
         let sellSum = fee + parseInt(sellAmount * sellPrecision, 10);
-        if (sellSum <= sellBalance) { // Sufficient balance in asset to pay fee
-            return feeAsset.get("id");
-        } else if (coreFee <= coreBalance && feeAsset.get("id") !== "1.3.0") { // Sufficient balance in core asset to pay fee
-            return "1.3.0";
+        if (feeAsset.get("id") === "1.3.0") {
+            if (coreFee <= coreBalance) {
+                return "1.3.0";
+            } else {
+                return null;
+            }
         } else {
-            return null; // Unable to pay fee in either asset
+            if (sellSum <= sellBalance) { // Sufficient balance in asset to pay fee
+                return feeAsset.get("id");
+            } else if (coreFee <= coreBalance && feeAsset.get("id") !== "1.3.0") { // Sufficient balance in core asset to pay fee
+                return "1.3.0";
+            } else {
+                return null; // Unable to pay fee in either asset
+            }
         }
     }
 
-    _createLimitOrderConfirm(buyAsset, sellAsset, buyAssetAmount, sellAssetAmount, sellBalance, coreBalance, type, e) {
+    _createLimitOrderConfirm(buyAsset, sellAsset, buyAssetAmount, sellAssetAmount, sellBalance, coreBalance, feeAsset, type, short = true, e) {
         e.preventDefault();
-        let feeID; // = "1.3.0";
 
         sellBalance = sellBalance ? parseInt(ChainStore.getObject(sellBalance).toJS().balance, 10) : 0;
         coreBalance = coreBalance ? parseInt(ChainStore.getObject(coreBalance).toJS().balance, 10) : 0;
@@ -317,13 +349,13 @@ class Exchange extends React.Component {
 
         // Convert fee to relevant asset fee and check if user has sufficient balance
         let feeAmount;
-        if (sellAsset.get("id") !== "1.3.0") {
-            feeAmount = this._getFee(sellAsset);
-        } else {
-            feeAmount = this._getFee();
-        }
+        // if (sellAsset.get("id") !== "1.3.0") {
+        //     feeAmount = this._getFee(sellAsset);
+        // } else {
+        feeAmount = this._getFee(feeAsset);
+        // }
             
-        feeID = this._verifyFee(feeAmount, sellAsset, sellAssetAmount, sellBalance, coreBalance);
+        let feeID = this._verifyFee(feeAmount, feeAsset, sellAsset, sellAssetAmount, sellBalance, coreBalance);
 
         if (!feeID) {
             return notify.addNotification({
@@ -350,7 +382,9 @@ class Exchange extends React.Component {
             }
         }
 
-        if ((sellAssetAmount * sellPrecision) > sellBalance) {
+        let isPredictionMarket = sellAsset.getIn(["bitasset", "is_prediction_market"]);
+
+        if ((sellAssetAmount * sellPrecision) > sellBalance && !isPredictionMarket) {
             return notify.addNotification({
                 message: "Insufficient funds to place order. Required: " + sellAssetAmount + " " + sellAsset.get("symbol"),
                 level: "error"
@@ -364,7 +398,35 @@ class Exchange extends React.Component {
             });
         }
 
+        if (type === "sell" && isPredictionMarket && short) {
+            return this._createPredictionShort(buyAsset, sellAsset, buyAssetAmount, sellAssetAmount, feeID);
+        }
         this._createLimitOrder(buyAsset, sellAsset, buyAssetAmount, sellAssetAmount, feeID);
+    }
+
+    _createLimitOrder(buyAsset, sellAsset, buyAssetAmount, sellAssetAmount, feeID) {
+        console.log("createLimitOrder:", buyAssetAmount, sellAssetAmount);
+        let expiration = new Date();
+        // TODO: Add selector for expiry
+        expiration.setYear(expiration.getFullYear() + 5);
+        MarketsActions.createLimitOrder(
+            this.props.currentAccount.get("id"),
+            utils.get_satoshi_amount(sellAssetAmount, sellAsset),
+            sellAsset,
+            utils.get_satoshi_amount(buyAssetAmount, buyAsset),
+            buyAsset,
+            expiration,
+            false, // fill or kill TODO: add fill or kill switch
+            feeID
+        ).then(result => {
+            if (result.error) {
+                if (result.error.message !== "wallet locked")
+                    notify.addNotification({
+                        message: "Unknown error. Failed to place order for " + buyAssetAmount + " " + buyAsset.symbol,
+                        level: "error"
+                    });
+            }
+        });
     }
 
     _forceBuy(buyAsset, sellAsset, buyAssetAmount, sellAssetAmount, value, sellBalance, coreBalance) {
@@ -430,24 +492,21 @@ class Exchange extends React.Component {
 
     _depthChartClick(base, quote, power, e) {
         e.preventDefault();
-        let value = this._limitByPrecision(e.xAxis[0].value / power, quote);
+        let value = market_utils.limitByPrecision(e.xAxis[0].value / power, quote);
         let buyPrice = this._getBuyPrice(e.xAxis[0].value / power);
         let sellPrice = this._getSellPrice(e.xAxis[0].value / power);
         let displayBuyPrice = this._getDisplayPrice("bid", buyPrice);
         let displaySellPrice = this._getDisplayPrice("ask", sellPrice);
-        // let buyPrice = this._buyPriceChanged(base, quote, {target: {value: value}});
 
         this.setState({
             depthLine: value,
             buyPrice: buyPrice,
             displayBuyPrice: displayBuyPrice,
-            buyTotal: this._limitByPrecision(this.getBuyTotal(buyPrice, this.state.buyAmount), base),
+            buyTotal: market_utils.limitByPrecision(this.getBuyTotal(buyPrice, this.state.buyAmount), base),
             sellPrice: sellPrice,
             displaySellPrice: displaySellPrice,
-            sellTotal: this._limitByPrecision(this.getSellTotal(sellPrice, this.state.sellAmount), base)
+            sellTotal: market_utils.limitByPrecision(this.getSellTotal(sellPrice, this.state.sellAmount), base)
         });
-
-        // this._sellPriceChanged(base, quote, {target: {value: value}});
     }
 
     _addZero(value) {
@@ -461,29 +520,38 @@ class Exchange extends React.Component {
         return value;
     }
 
-    _limitByPrecision(value, asset, floor = true) {
-        let assetPrecision = asset.toJS ? asset.get("precision") : asset.precision;
-        let valueString = value.toString();
-        let splitString = valueString.split(".");
-        if (splitString.length === 1 || splitString.length === 2 && splitString[1].length <= assetPrecision) {
-            return value;
-        }
-        let precision = utils.get_asset_precision(assetPrecision);
-        value = floor ? Math.floor(value * precision) / precision : Math.round(value * precision) / precision;
-        if (isNaN(value) || !isFinite(value)) {
-            return 0;
-        }
-        return value;
-    }
+    _
 
     _buyPriceChanged(base, quote, e) {
-        let amount = this._limitByPrecision(e.target.value, base);
+
+        let split = e.target.value.split(".");
+        if (split.length === 2 && split[1].length === SATOSHI + 1) {
+            return;
+        }
+
+        let amount = market_utils.limitByPrecision(e.target.value, {precision: SATOSHI});
         let price = this._getBuyPrice(amount);
 
         this.setState({
             buyPrice: price,
             displayBuyPrice: amount,
-            buyTotal: this._limitByPrecision(this.getBuyTotal(price, this.state.buyAmount), base),
+            buyTotal: market_utils.limitByPrecision(this.getBuyTotal(price, this.state.buyAmount), base),
+            depthLine: amount
+        });
+    }
+
+    _sellPriceChanged(base, quote, e) {
+        let split = e.target.value.split(".");
+        if (split.length === 2 && split[1].length === SATOSHI + 1) {
+            return;
+        }
+        let amount = market_utils.limitByPrecision(e.target.value, {precision: SATOSHI});
+        let price = this._getSellPrice(amount);
+
+        this.setState({
+            sellPrice: price,
+            displaySellPrice: amount,
+            sellTotal: market_utils.limitByPrecision(this.getSellTotal(price, this.state.sellAmount), base),
             depthLine: amount
         });
     }
@@ -491,25 +559,25 @@ class Exchange extends React.Component {
     _buyAmountChanged(base, quote, e) {
         let value = e.target.value;
         if (e.target.value.indexOf(".") !== e.target.value.length -1) {
-            value = this._limitByPrecision(e.target.value, quote);
+            value = market_utils.limitByPrecision(e.target.value, quote);
         }
 
         this.setState({
             buyAmount: this._addZero(value),
-            buyTotal: this._limitByPrecision(this.getBuyTotal(this.state.buyPrice, value), base)
+            buyTotal: market_utils.limitByPrecision(this.getBuyTotal(this.state.buyPrice, value), base)
         });
     }
 
     _buyTotalChanged(base, quote, e) {
         let value = e.target.value;
         if (e.target.value.indexOf(".") !== e.target.value.length -1) {
-            value = this._limitByPrecision(e.target.value, base);
+            value = market_utils.limitByPrecision(e.target.value, base);
         }
 
         let amount = this.getBuyAmount(this.state.buyPrice, value);
 
         this.setState({
-            buyAmount: this._limitByPrecision(amount, quote),
+            buyAmount: market_utils.limitByPrecision(amount, quote),
             buyTotal: this._addZero(value)
         });
     }
@@ -517,34 +585,22 @@ class Exchange extends React.Component {
     _sellAmountChanged(base, quote, e) {
         let value = e.target.value;
         if (e.target.value.indexOf(".") !== e.target.value.length -1) {
-            value = this._limitByPrecision(e.target.value, quote);
+            value = market_utils.limitByPrecision(e.target.value, quote);
         }
         this.setState({
             sellAmount: this._addZero(value),
-            sellTotal: this._limitByPrecision(this.getSellTotal(this.state.sellPrice, value), base)
-        });
-    }
-
-    _sellPriceChanged(base, quote, e) {
-        let amount = this._limitByPrecision(e.target.value, base);
-        let price = this._getSellPrice(amount);
-
-        this.setState({
-            sellPrice: price,
-            displaySellPrice: amount,
-            sellTotal: this._limitByPrecision(this.getSellTotal(price, this.state.sellAmount), base),
-            depthLine: amount
+            sellTotal: market_utils.limitByPrecision(this.getSellTotal(this.state.sellPrice, value), base)
         });
     }
 
     _sellTotalChanged(base, quote, e) {
         let value = e.target.value;
         if (e.target.value.indexOf(".") !== e.target.value.length -1) {
-            value = this._limitByPrecision(e.target.value, base);
+            value = market_utils.limitByPrecision(e.target.value, base);
         }
 
         this.setState({
-            sellAmount: this._limitByPrecision(this.getSellAmount(this.state.sellPrice, value), quote),
+            sellAmount: market_utils.limitByPrecision(this.getSellAmount(this.state.sellPrice, value), quote),
             sellTotal: this._addZero(value)
         });
     }
@@ -579,6 +635,7 @@ class Exchange extends React.Component {
     getBuyTotal(price, amount = 0) {
         let totalPrecision = utils.get_asset_precision(this.props.baseAsset.get("precision"));
         let satAmount = utils.get_satoshi_amount(amount, this.props.quoteAsset);
+
         return ((satAmount / price.base.amount) * price.quote.amount) / totalPrecision;
     }
 
@@ -604,9 +661,9 @@ class Exchange extends React.Component {
             let {buyTotal, buyAmount} = this.state;
 
             if (buyAmount) {
-                buyTotal = this._limitByPrecision(this.getBuyTotal(price, this.state.buyAmount), base);
+                buyTotal = market_utils.limitByPrecision(this.getBuyTotal(price, this.state.buyAmount), base);
             } else if (buyTotal) {
-                buyAmount = this._limitByPrecision(this.getBuyAmount(price, buyTotal), quote);
+                buyAmount = market_utils.limitByPrecision(this.getBuyAmount(price, buyTotal), quote);
             }
 
             this.setState({
@@ -621,9 +678,9 @@ class Exchange extends React.Component {
             let {sellTotal, sellAmount} = this.state;
 
             if (sellAmount) {
-                sellTotal = this._limitByPrecision(this.getSellTotal(price, this.state.sellAmount), base);
+                sellTotal = market_utils.limitByPrecision(this.getSellTotal(price, this.state.sellAmount), base);
             } else if (sellTotal) {
-                sellAmount = this._limitByPrecision(this.getSellAmount(price, sellTotal), quote);
+                sellAmount = market_utils.limitByPrecision(this.getSellAmount(price, sellTotal), quote);
             }
             this.setState({
                 sellPrice: price,
@@ -638,12 +695,13 @@ class Exchange extends React.Component {
 
     _orderbookClick(base, quote, type, order) {
         let precision = utils.get_asset_precision(quote.get("precision") + base.get("precision"));
-        if (type === "bid") {
-            let value = order.totalAmount.toString();
-            if (value.indexOf(".") !== value.length -1) {
-                value = this._limitByPrecision(order.totalAmount, quote);
-            }
+        
+        let value = order.amount.toString();
+        if (value.indexOf(".") !== value.length - 1) {
+            value = market_utils.limitByPrecision(order.amount, quote);
+        }
 
+        if (type === "bid") {
 
             let displaySellPrice = this._getDisplayPrice("ask", order.sell_price);
 
@@ -653,14 +711,10 @@ class Exchange extends React.Component {
                 displaySellPrice: displaySellPrice,
                 sellPrice: order.sell_price,
                 sellAmount: value,
-                sellTotal: this._limitByPrecision(total, base)
+                sellTotal: market_utils.limitByPrecision(total, base)
             });
 
         } else if (type === "ask") {
-            let value = order.totalAmount.toString();
-            if (value.indexOf(".") !== value.length -1) {
-                value = this._limitByPrecision(order.totalAmount, base);
-            }
 
             let displayBuyPrice = this._getDisplayPrice("bid", order.sell_price);
 
@@ -670,7 +724,7 @@ class Exchange extends React.Component {
                 displayBuyPrice: displayBuyPrice,
                 buyPrice: order.sell_price,
                 buyAmount: value,
-                buyTotal: this._limitByPrecision(total, base)
+                buyTotal: market_utils.limitByPrecision(total, base)
             });
         }
     }
@@ -709,18 +763,17 @@ class Exchange extends React.Component {
 
     _getDisplayPrice(type, priceObject) {
         let {quoteAsset, baseAsset} = this.props;
-        let precision =  Math.min(8, quoteAsset.get("precision") + baseAsset.get("precision"));
         let price;
 
         switch (type) {
             case "bid":
                 price = utils.get_asset_price(priceObject.quote.amount, baseAsset, priceObject.base.amount, quoteAsset);
-                price = this._limitByPrecision(this._addZero(price), {precision}, false);
+                price = market_utils.limitByPrecision(this._addZero(price), {precision: SATOSHI}, false);
                 return isNaN(price) ? 0 : price;
 
             case "ask":
                 price = utils.get_asset_price(priceObject.base.amount, baseAsset, priceObject.quote.amount, quoteAsset);
-                price = this._limitByPrecision(this._addZero(price), {precision}, false);
+                price = market_utils.limitByPrecision(this._addZero(price), {precision: SATOSHI}, false);
                 return isNaN(price) ? 0 : price;
 
             default:
@@ -753,7 +806,8 @@ class Exchange extends React.Component {
 
     _parseMarket() {
         let {bids, asks, calls, invertedCalls} = this.props;
-        let {showCallLimit} = this._getSettlementInfo;
+
+        let {showCallLimit} = this._getSettlementInfo();
         let combinedAsks, combinedBids, highestBid, lowestAsk;
 
         if (calls.length && invertedCalls) {
@@ -831,7 +885,7 @@ class Exchange extends React.Component {
                         }
                     }, null);
                     squeezePrice = settlementPrice / short_squeeze;
-                    showCallLimit = highestBid < lowestCallPrice && lowestCallPrice > squeezePrice;
+                    showCallLimit = lowestCallPrice > settlementPrice;
                 } else {
                     lowestAsk = asks.reduce((total, ask) => {
                         if (!total) {
@@ -842,7 +896,7 @@ class Exchange extends React.Component {
                     }, null);
 
                     squeezePrice = settlementPrice * short_squeeze;
-                    showCallLimit = lowestAsk > lowestCallPrice && lowestCallPrice < squeezePrice;
+                    showCallLimit = lowestCallPrice < settlementPrice;
                 }
             }
         }
@@ -878,6 +932,28 @@ class Exchange extends React.Component {
         SettingsActions.changeViewSetting({
             indicatorSettings: indicatorSettings
         });
+    }
+
+    onChangeFeeAsset(type, e) {
+        e.preventDefault();
+
+        if (type === "buy") {
+            this.setState({
+                preferCoreBuyFee: !this.state.preferCoreBuyFee
+            });
+
+            SettingsActions.changeViewSetting({
+                "preferCoreBuyFee": e.target.value === "1.3.0"
+            });
+        } else {
+            this.setState({
+                preferCoreSellFee: !this.state.preferCoreSellFee
+            });
+            
+            SettingsActions.changeViewSetting({
+                "preferCoreSellFee": e.target.value === "1.3.0"
+            });
+        }
     }
 
     render() {
@@ -923,7 +999,7 @@ class Exchange extends React.Component {
         let quoteIsBitAsset = quoteAsset.get("bitasset_data_id") ? true : false;
         let baseIsBitAsset = baseAsset.get("bitasset_data_id") ? true : false;
 
-        let {combinedAsks, combinedBids, spread, lowestAsk, highestBid} = this._parseMarket(showCallLimit);
+        let {combinedAsks, combinedBids, spread, lowestAsk, highestBid} = this._parseMarket();
 
         // Latest price
         if (activeMarketHistory.size) {
@@ -992,11 +1068,30 @@ class Exchange extends React.Component {
         let starClass = starredMarkets.has(marketID) ? "gold-star" : "grey-star";
 
         // Fees
-        let sellFee = utils.round_number(utils.get_asset_amount(this._getFee(quote), quote), quote);
-        let buyFee = utils.round_number(utils.get_asset_amount(this._getFee(base), base), base);
+        let coreAsset = ChainStore.getAsset("1.3.0");
+        if (!coreAsset) {
+            return null;
+        }
+
+        let sellFeeAsset = this.state.preferCoreSellFee ? coreAsset : quote !== coreAsset ? quote : base;
+        let sellFee = utils.round_number(utils.get_asset_amount(this._getFee(sellFeeAsset), sellFeeAsset), sellFeeAsset);
+        let sellFeeAssets = [coreAsset, quote === coreAsset ? base : quote];
+
+        let buyFeeAsset = this.state.preferCoreBuyFee ? coreAsset : base !== coreAsset ? base : quote;
+        let buyFee = utils.round_number(utils.get_asset_amount(this._getFee(buyFeeAsset), buyFeeAsset), buyFeeAsset);
+        let buyFeeAssets = [coreAsset, base === coreAsset ? quote : base];
 
         // Decimals
         let priceDecimals = Math.max(5, base ? base.get("precision") : 0);
+
+        let hasPrediction = base.getIn(["bitasset", "is_prediction_market"]) || quote.getIn(["bitasset", "is_prediction_market"]);
+
+        let description = null;
+
+        if (hasPrediction) {
+            description = quoteAsset.getIn(["options", "description"]);
+            description = assetUtils.parseDescription(description).main;
+        }
 
         return (
                 <div className="grid-block page-layout market-layout">
@@ -1028,13 +1123,24 @@ class Exchange extends React.Component {
                     </div>) : null}
 
                     {/* Center Column */}
-                    <div className={cnames("grid-block main-content vertical ps-container")} >
+                    <div className={cnames("grid-block main-content vertical no-overflow")} >
 
                         {/* Top bar with info */}
                         <div className="grid-block no-padding shrink overflow-visible top-bar" style={{minHeight: "67px"}}>
-                            <div className="grid-block overflow-visible">
+                            <div className="grid-block no-overflow">
                                 <div className="grid-block shrink" style={{borderRight: "1px solid grey"}}>
-                                    <span style={{paddingRight: 0}} onClick={this._addMarket.bind(this, quoteAsset.get("symbol"), baseAsset.get("symbol"))} className="market-symbol"><Icon className={starClass} name="fi-star"/></span><Link className="market-symbol" to={`/market/${baseSymbol}_${quoteSymbol}`}><span>{`${quoteSymbol} : ${baseSymbol}`}</span></Link>
+                                    <span style={{paddingRight: 0}} onClick={this._addMarket.bind(this, quoteAsset.get("symbol"), baseAsset.get("symbol"))} className="market-symbol">
+                                        <Icon className={starClass} name="fi-star"/>
+                                    </span>
+                                    {!hasPrediction ? (
+                                        <Link className="market-symbol" to={`/market/${baseSymbol}_${quoteSymbol}`}>
+                                            <span>{`${quoteSymbol} : ${baseSymbol}`}</span>
+                                        </Link>) : (
+                                        <a className="market-symbol">
+                                            <span>{`${quoteSymbol} : ${baseSymbol}`}</span>
+                                        </a>
+                                        )}
+
                                 </div>
                                 <div className="grid-block vertical">
                                     <div className="grid-block wrap" style={{borderBottom: "1px solid grey"}}>
@@ -1121,7 +1227,7 @@ class Exchange extends React.Component {
                                 </div>
                             </div>
                         </div>
-                        <div ref="center">
+                        <div className="grid-block vertical no-padding market-right-padding" id="CenterContent" ref="center">
                         {!this.state.showDepthChart ? (
                             <div className="grid-block shrink" id="market-charts" style={{marginTop: 0}}>
                                 {/* Price history chart */}
@@ -1140,7 +1246,26 @@ class Exchange extends React.Component {
                                     indicatorSettings={indicatorSettings}
                                     bucketSize={bucketSize}
                                     latest={latestPrice}
-                                />
+                                    verticalOrderbook={leftOrderBook}
+                                >
+                                    <div className="float-right">
+                                    <ul className="market-stats stats bottom-stats">
+                                        <li className="stat" style={{minHeight: "2rem"}}>
+                                            <span>
+                                                <span><Translate content="exchange.time" />:</span>
+                                                <span>{bucketOptions}</span>
+                                                <span></span>
+                                            </span>
+                                        </li>
+                                        <li className="stat clickable" onClick={this._onSelectIndicators.bind(this)}>
+                                            <div className="indicators">
+                                                <Translate content="header.settings" />
+                                            </div>
+                                        </li>
+                                    </ul>
+                                    </div>
+
+                                </PriceChart>
                                 <IndicatorModal
                                     ref="indicators"
                                     indicators={indicators}
@@ -1170,21 +1295,25 @@ class Exchange extends React.Component {
                                     plotLine={this.state.depthLine}
                                     settlementPrice={settlementPrice}
                                     spread={spread}
-                                    SQP={showCallLimit ? squeezePrice : null}
+                                    SQP={showCallLimit ? null : null}
                                     LCP={showCallLimit ? lowestCallPrice : null}
                                     leftOrderBook={leftOrderBook}
+                                    hasPrediction={hasPrediction}
+                                    noFrame={false}
+                                    verticalOrderbook={leftOrderBook}
                                 />
-
                             </div>)}
 
                         {/* Buy/Sell forms */}
 
                         {isNullAccount ? null : (
                             <div className="grid-block vertical shrink buy-sell">
-                            <div className="grid-block small-vertical medium-horizontal align-spaced" style={{ flexGrow: "0" }} >
+                            {hasPrediction ? <div className="grid-content no-overflow" style={{lineHeight: "1.2rem", paddingTop: 10}}>{description}</div> : null}
+                            
+                            <div className="grid-block small-vertical medium-horizontal align-spaced">
                                 {quote && base ?
                                 <BuySell
-                                    className={cnames("small-12 medium-5 no-padding", this.state.flipBuySell ? "order-3 sell-form" : "order-1 buy-form")}
+                                    className={cnames("small-12 medium-6 no-padding", this.state.flipBuySell ? "order-2 sell-form" : "order-1 buy-form")}
                                     type="bid"
                                     amount={buyAmount}
                                     price={displayBuyPrice}
@@ -1196,7 +1325,7 @@ class Exchange extends React.Component {
                                     setPrice={this._currentPriceClick.bind(this, base, quote)}
                                     totalChange={this._buyTotalChanged.bind(this, base, quote)}
                                     balance={baseBalance}
-                                    onSubmit={this._createLimitOrderConfirm.bind(this, quote, base, buyAmount, buyTotal, baseBalance, coreBalance, "buy")}
+                                    onSubmit={this._createLimitOrderConfirm.bind(this, quote, base, buyAmount, buyTotal, baseBalance, coreBalance, buyFeeAsset, "buy")}
                                     balancePrecision={base.get("precision")}
                                     quotePrecision={quote.get("precision")}
                                     totalPrecision={base.get("precision")}
@@ -1204,6 +1333,11 @@ class Exchange extends React.Component {
                                     currentPriceObject={lowestAsk.sell_price}
                                     account={currentAccount.get("name")}
                                     fee={buyFee}
+                                    feeAssets={buyFeeAssets}
+                                    feeAsset={buyFeeAsset}
+                                    onChangeFeeAsset={this.onChangeFeeAsset.bind(this, "buy")}
+                                    isPredictionMarket={base.getIn(["bitasset", "is_prediction_market"])}
+                                    onFlip={!this.state.flipBuySell ? this._flipBuySell.bind(this) : null}
                                 /> : null}
                                 <ConfirmOrderModal
                                     type="buy"
@@ -1212,12 +1346,9 @@ class Exchange extends React.Component {
                                     diff={buyDiff}
                                 />
 
-                                <div onClick={this._flipBuySell.bind(this)} className="grid-block vertical align-center text-center no-padding shrink order-2" style={{cursor: "pointer"}}>
-                                    <span style={{fontSize: "2rem"}}>&#8646;</span>
-                                </div>
                                 {quote && base ?
                                 <BuySell
-                                    className={cnames("small-12 medium-5 no-padding", this.state.flipBuySell ? "order-1 buy-form" : "order-3 sell-form")}
+                                    className={cnames("small-12 medium-6 no-padding", this.state.flipBuySell ? "order-1 buy-form" : "order-2 sell-form")}
                                     type="ask"
                                     amount={sellAmount}
                                     price={displaySellPrice}
@@ -1229,7 +1360,7 @@ class Exchange extends React.Component {
                                     setPrice={this._currentPriceClick.bind(this, base, quote)}
                                     totalChange={this._sellTotalChanged.bind(this, base, quote)}
                                     balance={quoteBalance}
-                                    onSubmit={this._createLimitOrderConfirm.bind(this, base, quote, sellTotal, sellAmount, quoteBalance, coreBalance, "sell")}
+                                    onSubmit={this._createLimitOrderConfirm.bind(this, base, quote, sellTotal, sellAmount, quoteBalance, coreBalance, sellFeeAsset, "sell")}
                                     balancePrecision={quote.get("precision")}
                                     quotePrecision={quote.get("precision")}
                                     totalPrecision={base.get("precision")}
@@ -1237,7 +1368,11 @@ class Exchange extends React.Component {
                                     currentPriceObject={highestBid.sell_price}
                                     account={currentAccount.get("name")}
                                     fee={sellFee}
-                                    ref="sell"
+                                    feeAssets={sellFeeAssets}
+                                    feeAsset={sellFeeAsset}
+                                    onChangeFeeAsset={this.onChangeFeeAsset.bind(this, "sell")}
+                                    isPredictionMarket={quote.getIn(["bitasset", "is_prediction_market"])}
+                                    onFlip={this.state.flipBuySell ? this._flipBuySell.bind(this) : null}
                                 /> : null}
                                 <ConfirmOrderModal
                                     type="sell"
@@ -1268,7 +1403,19 @@ class Exchange extends React.Component {
 
                         {isNullAccount ? null : (
                             <div className="grid-block no-overflow shrink no-padding">
-                            {limit_orders.size > 0 && base && quote ? (
+                                <MarketHistory
+                                    className="no-padding no-overflow"
+                                    headerStyle={{paddingTop: 0}}
+                                    history={activeMarketHistory}
+                                    myHistory={currentAccount.get("history")}
+                                    base={base}
+                                    quote={quote}
+                                    baseSymbol={baseSymbol}
+                                    quoteSymbol={quoteSymbol}
+                                    isNullAccount={isNullAccount}
+                                />
+
+                                {limit_orders.size > 0 && base && quote ? (
                                 <MyOpenOrders
                                     key="open_orders"
                                     orders={limit_orders}
@@ -1283,6 +1430,7 @@ class Exchange extends React.Component {
                         </div>)}
 
                         <div className="grid-block no-overflow shrink no-padding">
+
                             {settle_orders.size > 0 && base && quote &&
                             (base.get("id") === "1.3.0" || quote.get("id") === "1.3.0") ? (
                                 <OpenSettleOrders
@@ -1304,22 +1452,9 @@ class Exchange extends React.Component {
 
 
                     {/* Right Column - Market History */}
-                    <div className="grid-block shrink right-column no-overflow vertical" style={{paddingTop: 0, paddingRight: "0.5rem"}}>
+                    <div className="grid-block shrink right-column no-overflow vertical" style={{paddingTop: 0, minWidth: 358}}>
                         {/* Market History */}
-                        <div className="grid-block no-padding no-margin vertical"  style={{flex: "1 1 50vh", borderBottom: "1px solid grey"}}>
-                            <MarketHistory
-                                className="left-order-book no-padding no-overflow"
-                                headerStyle={{paddingTop: 0}}
-                                history={activeMarketHistory}
-                                myHistory={currentAccount.get("history")}
-                                base={base}
-                                quote={quote}
-                                baseSymbol={baseSymbol}
-                                quoteSymbol={quoteSymbol}
-                                isNullAccount={isNullAccount}
-                            />
-                        </div>
-                        <div className="grid-block no-padding no-margin vertical" style={{flex: "0 1 50vh"}}>
+                        <div className="grid-block no-padding no-margin vertical"  style={{borderBottom: "1px solid grey"}}>
                             <MyMarkets
                                 className="left-order-book no-padding no-overflow"
                                 headerStyle={{paddingTop: 0}}
@@ -1335,17 +1470,47 @@ class Exchange extends React.Component {
                                 current={marketID}
                             />
                         </div>
+                        <div className="grid-block no-padding no-margin vertical shrink">
+                            <DepthHighChart
+                                    orders={limit_orders}
+                                    showCallLimit={showCallLimit}
+                                    call_orders={call_orders}
+                                    flat_asks={flat_asks}
+                                    flat_bids={flat_bids}
+                                    flat_calls={ showCallLimit ? flat_calls : []}
+                                    settles={settle_orders}
+                                    invertedCalls={invertedCalls}
+                                    totalBids={totalBids}
+                                    totalCalls={showCallLimit ? totalCalls : 0}
+                                    base={base}
+                                    quote={quote}
+                                    baseSymbol={baseSymbol}
+                                    quoteSymbol={quoteSymbol}
+                                    height={200}
+                                    onClick={this._depthChartClick.bind(this, base, quote)}
+                                    plotLine={this.state.depthLine}
+                                    settlementPrice={settlementPrice}
+                                    spread={spread}
+                                    SQP={showCallLimit ? squeezePrice : null}
+                                    LCP={showCallLimit ? lowestCallPrice : null}
+                                    leftOrderBook={leftOrderBook}
+                                    hasPrediction={hasPrediction}
+                                    noText={true}
+                                />                                
+                        </div>
                     </div>
-                    {!isNullAccount && quoteIsBitAsset ?
+                    {!isNullAccount && quoteIsBitAsset  ?
                         <BorrowModal
                             ref="borrowQuote"
                             quote_asset={quoteAsset.get("id")}
+                            backing_asset={quoteAsset.getIn(["bitasset", "options", "short_backing_asset"])}
                             account={currentAccount}
                          /> : null}
-                    {!isNullAccount && baseIsBitAsset ?
+                    {!isNullAccount && baseIsBitAsset ? 
                         <BorrowModal
                             ref="borrowBase"
                             quote_asset={baseAsset.get("id")}
+                            backing_asset={baseAsset.getIn(["bitasset", "options", "short_backing_asset"])}
                             account={currentAccount}
                         /> : null}
                 {/* End of Second Vertical Block */}
